requirements.txt
anyio==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Markdown==3.5.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic_core==2.16.2
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
